\name{iGREX-package}
\alias{iGREX-package}
\title{
IGREX for quantifying the impact of genetically regulated expression on phenotypes
}
\description{
This package provides functions for fitting IGREX, a statistical model for quantifying the impact of genetically regulated expression on phenotypes.
}
\usage{
iGREX-package
}
\details{
\tabular{ll}{
Package: \tab LPM\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2018-07-10\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}

The main fucntions of this package are \code{iGREX} and \code{iGREXs}, which fit IGREX-i and IGREX-s model, respectively.
}
\references{

}
\author{
  \packageAuthor{iGREX}

  Maintainer: \packageMaintainer{iGREX}
}
\note{
Please refer to the package vignette for details.
}


\seealso{
\code{\link{iGREX}}, \code{\link{iGREXs}}
}
\keyword{package}
